chore: Rename namespace to app.icampuspass
Rename namespace from com.itocc.icampuspass to app.icampuspass

Bug: #49
