+++ composeApp/build.gradle.kts
@@ -37,10 +37,10 @@ kotlin {
 }
 
 android {
-    namespace = "com.itocc.icampuspass"
+    namespace = "app.icampuspass"
     compileSdk = libs.versions.android.compileSdk.get().toInt()
     defaultConfig {
-        applicationId = "com.itocc.icampuspass"
+        applicationId = "app.icampuspass"
         minSdk = libs.versions.android.minSdk.get().toInt()
         targetSdk = libs.versions.android.targetSdk.get().toInt()
         versionCode = 1
