--- composeApp/src/androidMain/kotlin/com/itocc/icampuspass/AndroidApp.kt
@@ -1,17 +1,0 @@
-package com.itocc.icampuspass
-
-import android.app.Application
-import org.koin.dsl.module
-
-class AndroidApp: Application() {
-    override fun onCreate() {
-        super.onCreate()
-
-        initKoin(
-            androidContext = this@AndroidApp,
-            extraModules = listOf(
-                module {}
-            )
-        )
-    }
-}
