+++ composeApp/src/androidUnitTest/kotlin/app/icampuspass/ComposeAppAndroidUnitTest.kt
@@ -1,0 +1,12 @@
+package app.icampuspass
+
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class ComposeAppAndroidUnitTest {
+
+    @Test
+    fun example() {
+        assertEquals(expected = 3, actual = 1 + 2)
+    }
+}
