--- composeApp/src/androidUnitTest/kotlin/com/itocc/icampuspass/ComposeAppAndroidUnitTest.kt
@@ -1,12 +1,0 @@
-package com.itocc.icampuspass
-
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class ComposeAppAndroidUnitTest {
-
-    @Test
-    fun example() {
-        assertEquals(expected = 3, actual = 1 + 2)
-    }
-}
