+++ iosApp/Configuration/Config.xcconfig
@@ -1,5 +1,5 @@
 TEAM_ID=
-BUNDLE_ID=com.itocc.icampuspass
+BUNDLE_ID=app.icampuspass
 PRODUCT_NAME=iCampusPass
 CURRENT_PROJECT_VERSION=1
 MARKETING_VERSION=1.0
