+++ shared/build.gradle.kts
@@ -65,7 +65,7 @@ kotlin {
 }
 
 android {
-    namespace = "com.itocc.icampuspass.shared"
+    namespace = "app.icampuspass.shared"
     compileSdk = libs.versions.android.compileSdk.get().toInt()
     compileOptions {
         sourceCompatibility = JavaVersion.VERSION_17
@@ -79,7 +79,7 @@ android {
 sqldelight {
     databases {
         create(name = "Database") {
-            packageName.set("com.itocc.icampuspass.models.database")
+            packageName.set("app.icampuspass.models.database")
             generateAsync.set(true)
         }
     }
