+++ shared/src/androidMain/kotlin/app/icampuspass/Platform.android.kt
@@ -1,0 +1,9 @@
+package app.icampuspass
+
+import android.os.Build
+
+class AndroidPlatform: Platform {
+    override val name: String = "Android ${Build.VERSION.SDK_INT}"
+}
+
+actual fun getPlatform(): Platform = AndroidPlatform()
