--- shared/src/androidMain/kotlin/com/itocc/icampuspass/Platform.android.kt
@@ -1,9 +1,0 @@
-package com.itocc.icampuspass
-
-import android.os.Build
-
-class AndroidPlatform: Platform {
-    override val name: String = "Android ${Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()
