--- shared/src/androidMain/kotlin/com/itocc/icampuspass/models/database/DriverFactory.android.kt
@@ -1,18 +1,0 @@
-package com.itocc.icampuspass.models.database
-
-import android.content.Context
-import app.cash.sqldelight.async.coroutines.synchronous
-import app.cash.sqldelight.db.SqlDriver
-import app.cash.sqldelight.driver.android.AndroidSqliteDriver
-
-actual class DriverFactory(
-    private val context: Context
-) {
-    actual fun createDriver(): SqlDriver {
-        return AndroidSqliteDriver(
-            schema = Database.Schema.synchronous(),
-            context = context,
-            name = DatabaseHelper.Companion.FILE_NAME
-        )
-    }
-}
