+++ shared/src/commonMain/kotlin/app/icampuspass/Greeting.kt
@@ -1,0 +1,9 @@
+package app.icampuspass
+
+class Greeting {
+    private val platform = getPlatform()
+
+    fun greet(): String {
+        return "Hello, ${platform.name}!"
+    }
+}
