+++ shared/src/commonMain/kotlin/app/icampuspass/Koin.kt
@@ -1,0 +1,40 @@
+package app.icampuspass
+
+import app.icampuspass.models.database.DatabaseHelper
+import io.ktor.client.HttpClient
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.client.plugins.cookies.HttpCookies
+import io.ktor.http.ContentType
+import io.ktor.serialization.kotlinx.json.json
+import kotlinx.serialization.json.Json
+import org.koin.core.context.startKoin
+import org.koin.core.module.Module
+import org.koin.dsl.module
+
+val commonModule: Module = module {
+    single<AppRepository> {
+        AppRepository().apply {
+            init()
+        }
+    }
+
+    single<DatabaseHelper> {
+        DatabaseHelper(driverFactory = get())
+    }
+}
+
+expect val platformModule: Module
+
+fun initKoin(
+    extraModules: List<Module>
+) {
+    startKoin {
+        modules(
+            commonModule,
+            platformModule,
+            *extraModules.toTypedArray(),
+        )
+    }
+}
+
+fun initKoin() = initKoin(extraModules = emptyList())
