+++ shared/src/commonMain/kotlin/app/icampuspass/Platform.kt
@@ -1,0 +1,7 @@
+package app.icampuspass
+
+interface Platform {
+    val name: String
+}
+
+expect fun getPlatform(): Platform
