+++ shared/src/commonMain/kotlin/app/icampuspass/models/database/DriverFactory.kt
@@ -1,0 +1,7 @@
+package app.icampuspass.models.database
+
+import app.cash.sqldelight.db.SqlDriver
+
+expect class DriverFactory {
+    fun createDriver(): SqlDriver
+}
