--- shared/src/commonMain/kotlin/com/itocc/icampuspass/AppRepository.kt
@@ -1,13 +1,0 @@
-package com.itocc.icampuspass
-
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.SupervisorJob
-import kotlinx.coroutines.launch
-
-class AppRepository() {
-    private val scope: CoroutineScope = CoroutineScope(context = SupervisorJob())
-
-    fun init() {
-        scope.launch {}
-    }
-}
