--- shared/src/commonMain/kotlin/com/itocc/icampuspass/Greeting.kt
@@ -1,9 +1,0 @@
-package com.itocc.icampuspass
-
-class Greeting {
-    private val platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
