--- shared/src/commonMain/kotlin/com/itocc/icampuspass/Platform.kt
@@ -1,7 +1,0 @@
-package com.itocc.icampuspass
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
