--- shared/src/commonMain/kotlin/com/itocc/icampuspass/models/database/DriverFactory.kt
@@ -1,7 +1,0 @@
-package com.itocc.icampuspass.models.database
-
-import app.cash.sqldelight.db.SqlDriver
-
-expect class DriverFactory {
-    fun createDriver(): SqlDriver
-}
