--- shared/src/commonTest/kotlin/com/itocc/icampuspass/SharedCommonTest.kt
@@ -1,12 +1,0 @@
-package com.itocc.icampuspass
-
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class SharedCommonTest {
-
-    @Test
-    fun example() {
-        assertEquals(expected = 3, actual = 1 + 2)
-    }
-}
