+++ shared/src/iosMain/kotlin/app/icampuspass/KoinDependencies.kt
@@ -1,0 +1,8 @@
+package app.icampuspass
+
+import org.koin.core.component.KoinComponent
+import org.koin.core.component.inject
+
+class KoinDependencies: KoinComponent {
+    val appRepository: AppRepository by inject()
+}
