+++ shared/src/iosMain/kotlin/app/icampuspass/Platform.ios.kt
@@ -1,0 +1,9 @@
+package app.icampuspass
+
+import platform.UIKit.UIDevice
+
+class IOSPlatform: Platform {
+    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
+}
+
+actual fun getPlatform(): Platform = IOSPlatform()
