+++ shared/src/iosMain/kotlin/app/icampuspass/models/database/DriverFactory.ios.kt
@@ -1,0 +1,14 @@
+package app.icampuspass.models.database
+
+import app.cash.sqldelight.async.coroutines.synchronous
+import app.cash.sqldelight.db.SqlDriver
+import app.cash.sqldelight.driver.native.NativeSqliteDriver
+
+actual class DriverFactory {
+    actual fun createDriver(): SqlDriver {
+        return NativeSqliteDriver(
+            schema = Database.Schema.synchronous(),
+            name = DatabaseHelper.Companion.FILE_NAME
+        )
+    }
+}
