--- shared/src/iosMain/kotlin/com/itocc/icampuspass/KoinDependencies.kt
@@ -1,8 +1,0 @@
-package com.itocc.icampuspass
-
-import org.koin.core.component.KoinComponent
-import org.koin.core.component.inject
-
-class KoinDependencies: KoinComponent {
-    val appRepository: AppRepository by inject()
-}
